# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BLASTERR social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clips:
    get:
      operationId: getMyClips
      tags: [clips]
      summary: Get the current user's BLASTR Clips
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, DRAFT, QUEUED, PROCESSING, COMPLETED, FAILED, SHARED]
            default: all
      responses:
        "200":
          description: Clip library
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Clip" }
    post:
      operationId: createClip
      tags: [clips]
      summary: Create a BLASTR Clip render job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClipInput"
      responses:
        "201":
          description: Queued clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clip"
        "403":
          description: Clip creation is not permitted
  /clips/{id}:
    get:
      operationId: getClip
      tags: [clips]
      summary: Get a BLASTR Clip
      parameters:
        - $ref: "#/components/parameters/ClipId"
      responses:
        "200":
          description: Clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clip"
        "404":
          description: Clip not found
    patch:
      operationId: updateClip
      tags: [clips]
      summary: Update a draft or failed BLASTR Clip
      parameters:
        - $ref: "#/components/parameters/ClipId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClipUpdate"
      responses:
        "200":
          description: Updated clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clip"
    delete:
      operationId: deleteClip
      tags: [clips]
      summary: Delete an owned BLASTR Clip
      parameters:
        - $ref: "#/components/parameters/ClipId"
      responses:
        "204":
          description: Deleted
  /clips/{id}/retry:
    post:
      operationId: retryClip
      tags: [clips]
      summary: Retry a failed BLASTR Clip render
      parameters:
        - $ref: "#/components/parameters/ClipId"
      responses:
        "200":
          description: Requeued clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clip"
  /clips/{id}/share:
    post:
      operationId: shareClip
      tags: [clips]
      summary: Create a share link for a completed BLASTR Clip
      parameters:
        - $ref: "#/components/parameters/ClipId"
      responses:
        "200":
          description: Share link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
...
  /admin/clips:
    get:
      operationId: getAdminClips
      tags: [admin, clips]
      summary: Get the protected BLASTR Media Center overview
      responses:
        "200":
          description: Media Center overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClipsOverview"
        "403":
          description: Admin access required
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    ClipId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, avatarUrl, bio, location, city, state, followers, following, blastCount, joinedAt]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        bio: { type: string }
        location: { type: string }
        city: { type: string }
        state: { type: string }
        coverUrl: { type: string }
        followers: { type: number }
        following: { type: number }
        blastCount: { type: number }
        joinedAt: { type: string }
        isFollowing: { type: boolean }
...
    Target:
      type: object
      required: [id, name, slug, type, location, blastCount, imageUrl, description]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        type:
          type: string
          enum: [person, business, place, product, entertainment, sports, gaming, other]
        location: { type: string }
        blastCount: { type: number }
        imageUrl: { type: string }
        description: { type: string }
...
    Blast:
      type: object
      required: [id, content, createdAt, author, target, location, mediaUrl, reactions, commentCount, shareCount, viewCount, isBookmarked]
      properties:
        id: { type: string }
        content: { type: string }
        createdAt: { type: string }
        author: { $ref: "#/components/schemas/User" }
        target: { $ref: "#/components/schemas/Target" }
        location: { type: string }
        mediaUrl: { type: string }
        mediaType:
          type: [string, "null"]
          enum: [image, video, null]
        reactions: { $ref: "#/components/schemas/ReactionSummary" }
        commentCount: { type: number }
        shareCount: { type: number }
        viewCount: { type: number }
        isBookmarked: { type: boolean }
        isBlastBack: { type: boolean }
        originalBlastId:
          type: [string, "null"]
...
    ReactionSummary:
      type: object
      required: [blast, facts, cap, funny, watching, currentUserReaction]
      properties:
        blast: { type: number }
        facts: { type: number }
        cap: { type: number }
        funny: { type: number }
        watching: { type: number }
        currentUserReaction:
          type: [string, "null"]
          enum: [blast, facts, cap, funny, watching, null]
...
    Clip:
      type: object
      required: [id, blastId, creatorId, style, title, description, duration, aspectRatio, renderStatus, videoUrl, thumbnailUrl, createdAt, updatedAt, blast, creator, settings]
      properties:
        id: { type: string }
        blastId: { type: string }
        creatorId: { type: string }
        style:
          type: string
          enum: [BREAKING, FUNNY, DRAMATIC, RECEIPTS, DEBATE, STORY]
        title: { type: string }
        description: { type: string }
        duration: { type: number }
        aspectRatio: { type: string, enum: ["9:16"] }
        renderStatus:
          type: string
          enum: [DRAFT, QUEUED, PROCESSING, COMPLETED, FAILED, SHARED]
        videoUrl:
          type: [string, "null"]
        thumbnailUrl:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
        createdAt: { type: string }
        updatedAt: { type: string }
        blast: { $ref: "#/components/schemas/Blast" }
        creator: { $ref: "#/components/schemas/User" }
        settings: { $ref: "#/components/schemas/ClipSettings" }
    ClipInput:
      type: object
      required: [blastId, style]
      properties:
        blastId: { type: string }
        style:
          type: string
          enum: [BREAKING, FUNNY, DRAMATIC, RECEIPTS, DEBATE, STORY]
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        captionStyle:
          type: string
          enum: [bold, minimal, breaking, meme, cinematic, story]
        captionPosition:
          type: string
          enum: [top, center, bottom]
        captionAnimation:
          type: string
          enum: [pop, slide, fade, word-by-word, highlight]
        background: { type: string, maxLength: 80 }
        duration:
          type: number
          enum: [15, 30, 45, 60]
        brandingStyle:
          type: string
          enum: [minimal, full]
        ctaText: { type: string, maxLength: 120 }
    ClipUpdate:
      type: object
      properties:
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        captionStyle: { type: string }
        captionPosition: { type: string }
        captionAnimation: { type: string }
        background: { type: string }
        duration: { type: number, enum: [15, 30, 45, 60] }
        brandingStyle: { type: string }
        ctaText: { type: string, maxLength: 120 }
    ClipSettings:
      type: object
      required: [captionStyle, captionPosition, captionAnimation, background, brandingStyle, ctaText]
      properties:
        captionStyle: { type: string }
        captionPosition: { type: string }
        captionAnimation: { type: string }
        background: { type: string }
        brandingStyle: { type: string }
        ctaText: { type: string }
    ShareLink:
      type: object
      required: [url]
      properties:
        url: { type: string }
    AdminClipsOverview:
      type: object
      required: [clipCount, completedCount, failedCount, activeRenders, clips]
      properties:
        clipCount: { type: number }
        completedCount: { type: number }
        failedCount: { type: number }
        activeRenders: { type: number }
        clips:
          type: array
          items: { $ref: "#/components/schemas/Clip" }
...